[outline] Widen origin negation
Synthetic slant translates outlines by the negated glyph origin.
Convert the full-range position to the float coordinate domain before
negating it, avoiding signed integer overflow at INT_MIN.

Tested with: meson test -C build --print-errorlogs

+++ src/hb-font.hh
@@ -738,7 +738,7 @@ struct hb_font_t
     {
       hb_position_t xo = 0, yo = 0;
       get_glyph_h_origin (glyph, &xo, &yo, false);
-      outline.translate (-xo, -yo);
+      outline.translate (-(float) xo, -(float) yo);
       outline.slant (slant_xy);
       outline.translate (xo, yo);
     }
